Don't panic when a non-nil error happens
A non-nil error coming back from the SSH command means the command being
run has returned a non-zero exit code. This doesn't necessarily means
the output is invalid. In this case, it was the output of a "systemctl
status" command for a service that was not running. In this case, the
exit code is 3 and the err structure returned from the SSH command
contains this exit code. The output of systemctl that shows the status
of the service is still returned in the output, so can still be used.