Update service test to handle preset/vendor preset
There are two possible strings that can be found in systemctl output
indicating whether or not the unit starts automatically after
installation - "preset" and "vendor preset". I'm not currently sure
about the difference between the two, but currently I'm assuming
"preset" is found on pre-installed units and "vendor preset" is for
user-installed units.