give resource queries more responsibility for parsing
Previously, the logic for parsing the output of a command was split into
two parts:
- parseOutput
- setValues

parseOutput would apply the regex to the output of the command and
setValues would take the values of the capture groups and assign them to
the fields of the type. parseOutput was a generic function so it would
be used for all resources.

This resulted in the following problems:
- all resources must be parsed using a regex. For simple command outputs
  like that for the user, this was overkill, and for complex command
  outputs, like that of systemctl, the regex would need to become too
  cumbersome to capture all possible states of the output.
- writing and debugging these regex string was the most time consuming
  part of adding a new resource, perhaps accounting for more time on its
  own than all the time taken to write the rest of the code for the
  resource combined, including tests.
- the regex strings were harder to read than go code, making them harder
  to maintain in the long run.

The solution to this has been to replace the two functions with a single
non-generic parseOutput attached to each resource query. This results
in:
- the SshClient having less responsibility for the processing of the
  command output, instead taking the raw string from the connection and
  passing it back to the resource query for processing, reducing the
  complexity of this type.
- more flexibility in how each resource is created from the command
  output, since each resource query can parse the output in whatever it
  sees fit without any influence from the others