Ignore error returned from RunCommand
This error doesn't contain any useful information for the client -
that's found in stderr which, in this case, is the output in the event
of an error.

For example, if querying a file that doesn't exist, err will say
"process exited with status 1" while output will say "no such file or
directory".

The err isn't useful here and I can't think of any immediate use for it,
so I'm going to ignore it for now and I can reintroduce it later if I
find a use for it.